Updating memory usage
